leaf53: split compile_v0 into input/stats/trace modules
